refactor: extract bench run metadata helpers
